+++ lib/net/gateway-web3.dart
@@ -14,11 +14,11 @@ class Web3Gateway {
   final String alternateEnvironment;
   Web3Client _client;
 
-  static String _entityResolverAddress; // Lazy loaded
-  static String _processAddress; // Lazy loaded
+  String _entityResolverAddress; // Lazy loaded
+  String _processAddress; // Lazy loaded
 
-  static DeployedContract _entityResolverInstance; // Loaded on init()
-  static DeployedContract _processInstance; // Loaded on init()
+  DeployedContract _entityResolverInstance; // Loaded on init()
+  DeployedContract _processInstance; // Loaded on init()
 
   String get uri => _gatewayUri;
 
@@ -66,23 +66,22 @@ class Web3Gateway {
     // Check contract address availability
     if (_entityResolverAddress is! String ||
         _entityResolverAddress.length == 0) {
-      Web3Gateway._entityResolverAddress =
-          await Web3Gateway.resolveEntityResolverDomain(this._gatewayUri,
-              alternateEnvironment: alternateEnvironment ?? false);
+      _entityResolverAddress = await Web3Gateway.resolveEntityResolverDomain(
+          this._gatewayUri,
+          alternateEnvironment: alternateEnvironment ?? "");
     }
     if (_processAddress is! String || _processAddress.length == 0) {
-      Web3Gateway._processAddress = await Web3Gateway.resolveProcessDomain(
-          this._gatewayUri,
+      _processAddress = await Web3Gateway.resolveProcessDomain(this._gatewayUri,
           alternateEnvironment: alternateEnvironment ?? "");
     }
 
     // Define contract instances
     _entityResolverInstance = DeployedContract(
         EnsPublicResolverContract.contractAbi,
-        EthereumAddress.fromHex(Web3Gateway._entityResolverAddress));
+        EthereumAddress.fromHex(_entityResolverAddress));
 
-    _processInstance = DeployedContract(ProcessContract.contractAbi,
-        EthereumAddress.fromHex(Web3Gateway._processAddress));
+    _processInstance = DeployedContract(
+        ProcessContract.contractAbi, EthereumAddress.fromHex(_processAddress));
   }
 
   void dispose() {
@@ -98,10 +97,10 @@ class Web3Gateway {
     DeployedContract contractInstance;
     switch (contractEnum) {
       case ContractEnum.EntityResolver:
-        contractInstance = Web3Gateway._entityResolverInstance;
+        contractInstance = _entityResolverInstance;
         break;
       case ContractEnum.Process:
-        contractInstance = Web3Gateway._processInstance;
+        contractInstance = _processInstance;
         break;
       default:
         throw Exception("Invalid contract enum value");
@@ -140,10 +139,10 @@ class Web3Gateway {
     DeployedContract contractInstance;
     switch (contractEnum) {
       case ContractEnum.EntityResolver:
-        contractInstance = Web3Gateway._entityResolverInstance;
+        contractInstance = _entityResolverInstance;
         break;
       case ContractEnum.Process:
-        contractInstance = Web3Gateway._processInstance;
+        contractInstance = _processInstance;
         break;
       default:
         throw Exception("Invalid contract enum value");
@@ -171,7 +170,7 @@ class Web3Gateway {
     return resolveName(domain, gatewayUri).then((address) {
       if (address is! String)
         throw Exception(
-            "The domain $ENS_PUBLIC_RESOLVER_DOMAIN does not resolve using $gatewayUri");
+            "The domain $domain does not resolve using $gatewayUri");
 
       return address;
     });
@@ -186,7 +185,7 @@ class Web3Gateway {
     return resolveName(domain, gatewayUri).then((address) {
       if (address is! String)
         throw Exception(
-            "The domain $PROCESS_DOMAIN does not resolve using $gatewayUri");
+            "The domain $domain does not resolve using $gatewayUri");
 
       return address;
     });