+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Hash entity address for ENS call
 - Convert ProcessData ints to BigInts
 - Ensure ProcessData parses upon declaration
+- Enable alternate ENS/Process contract domains
 
 ## 0.19.0
 - Update process contract wrappers to smart contracts v2